# openapi/openapi.yaml
openapi: 3.0.2
info:
    title: Hourly API
    version: '0.1'
    description: Provides interaction with Hourly web app.
    contact:
        name: Hourly
        url: 'http://www.hourly.com'
        email: test@test.com
servers:
    -
        url: 'http://localhost:8080/api/v0'
        description: Represents the local server.
    -
        url: 'http://hourly-timesheet-dev.herokuapp.com/api/v0'
        description: The staging API environment.
    -
        url: 'http://hourly-timesheet-prod.herokuapp.com/api/v0'
        description: The prod API environment.
paths:
...
    /events:
        get:
            tags:
                - Events
            parameters:
                -
                    style: form
                    explode: true
                    name: q
                    description: default query parameter
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: include_totals
                    description: include the total count in the response
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: page
                    description: The page to query from. This will take precedence over a limit and offset.
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: fields
                    description: 'Fields to exclude {"measurements": 0} or only include {"measurements": 1}'
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: sort
                    description: >-
                        comma-separated list of fields to define the sort order. To indicate sorting
                        direction, may be prefixed with + (ascending) or - (descending), e.g.
                        /sales-orders?sort=+id
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: offset
                    description: numeric offset of the first element on a page
                    schema:
                        type: integer
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: limit
                    description: client suggested limit to restrict the number of entries on a page
                    schema:
                        type: integer
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: lang
                    description: 'language code to prefer for document elements (en-US, fr ...) page'
                    schema:
                        type: string
                    in: query
                    required: false
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventListResponse'
                    description: Successully retrieved list of events.
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: list_events
            summary: List Events
            x-openapi-router-controller: domains.events.controllers.get.controller
        post:
            requestBody:
                description: Add Event
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Event'
                required: true
                x-body-name: event
            tags:
                - Events
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AddResponse'
                    description: event added
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: add_event
            summary: Add Event
            x-openapi-router-controller: domains.events.controllers.post.controller
    '/events/{id}':
        get:
            tags:
                - Events
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '200':
                    description: Successfully retrieved Event
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: get_event
            summary: Get Event
            x-openapi-router-controller: domains.events.controllers.get.controller
        delete:
            tags:
                - Events
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '204':
                    description: Event deleted
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: delete_event
            summary: Delete Event
            x-openapi-router-controller: domains.events.controllers.delete.controller
        patch:
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: '#/components/schemas/PatchDocument'
                required: true
                x-body-name: patch_document_list
            tags:
                - Events
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '204':
                    description: Event patched
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: patch_event
            summary: Patch Event
            x-openapi-router-controller: domains.events.controllers.patch.controller
...
    /events/search:
        post:
            requestBody:
                description: Search Event
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EventSearch'
                required: true
                x-body-name: search_query
            tags:
                - Events
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AddResponse'
                    description: results
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: search_events
            summary: Search Events
            description: Searches events within the user's company.
            x-openapi-router-controller: domains.events.controllers.post.controller
components:
    schemas:
        AddResponse:
            title: AddResponse
            description: Response to adding a new row.
            required:
                - id
            type: object
            properties:
                id:
                    title: id
                    type: integer
            example:
                id: id
...
        ErrorListResponse:
            title: Root Type for ErrorListResponse
            description: A list of error messages.
            required:
                - errors
            type: object
            properties:
                errors:
                    description: List of all error messages.
                    type: array
                    items:
                        title: ErrorMessage
                        type: object
            example: >-
                {'detail': 'The server encountered an unexpected error. Please try again later.','status':
                500,'suggestion': 'Please wait 5 minutes and re-attempt your request.'}
        Event:
            title: Event
            description: Represents a type of event that can be created.
            required:
                - name
                - description
                - agreed_price
                - start_datetime
                - package_id
                - questions
            type: object
            properties:
                id:
                    description: The identifier for the event.
                    type: integer
                name:
                    description: name
                    type: string
                description:
                    description: Block of text describing the event.
                    type: string
                agreed_price:
                    description: The agreed upon price.
                    type: number
                start_datetime:
                    description: The start datetime.
                    type: string
                end_datetime:
                    description: The end datetime.
                    type: string
                package_id:
                    description: The ID of the package associated with this event.
                    type: integer
                company_id:
                    description: The ID of the company associated to the event.
                    type: integer
                employee_id:
                    description: The ID of the employee servicing the event.
                    type: integer
                questions:
                    description: The questions to include in the event form.
                    type: array
                    items:
                        $ref: '#/components/schemas/PackageQuestion'
            example: >-
                {'name': 'Example', 'description': 'Example description', 'agreed_price': 'price',
                'start_datetime': '203919174', 'end_datetime': '239829328', 'package_id': 1, 'employee_id': 1,
                'questions': [{'title': 'Example', 'value': '', 'values': ['Test'], 'data_type':
                'multiselect'}] }
...
        PackageQuestion:
            title: PackageQuestion
            description: Represents a question type to be included in a package.
            required:
                - title
                - data_type
            type: object
            properties:
                title:
                    description: Text that outlines the question.
                    type: string
                value:
                    description: The value stored by the question.
                    type: string
                values:
                    description: 'The potential values, if any, of the question.'
                    type: array
                    items:
                        type: string
                data_type:
                    description: The type of the question.
                    type: string
            example: '{''title'': ''Example'', ''value'': '''', ''values'': [''Test''], ''data_type'': ''multiselect''}'
        PatchDocument:
            title: PatchDocument
            description: A JSONPatch document as defined by RFC 6902.
            required:
                - op
                - path
            type: object
            properties:
                op:
                    description: The operation to be performed
                    enum:
                        - add
                        - remove
                        - replace
                        - move
                        - copy
                        - test
                    type: string
                path:
                    description: A JSON-Pointer
                    type: string
                value:
                    description: The value to be used within the operations.
                    type: string
                fromPath:
                    description: A string containing a JSON Pointer value.
                    type: string
            example: |-
                {
                    "op": "replace",
                    "path": "/name",
                    "value": "Sample3"
                }
...
        EventListResponse:
            title: ListResponse
            description: A response that contains data.
            required:
                - data
                - message
                - status
            type: object
            properties:
                data:
                    title: data
                    type: array
                    items:
                        $ref: '#/components/schemas/Event'
                message:
                    title: message
                    type: string
                status:
                    title: status
                    type: integer
            example:
                message: Success.
                status: 200
                data:
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
...
        EventSearch:
            title: Root Type for EventSearch
            description: Represents an entity used to search events within the user's company.
            required:
                - from_date
                - to_date
            type: object
            properties:
                from_date:
                    description: The earliest date a result can take place at.
                    type: string
                to_date:
                    description: Represents the latest date a resultant event can be.
                    type: string
                package_name:
                    description: The name of the package.
                    type: string
            example:
                from_date: 01/01/2020
                to_date: 02/20/2022
                package_name: dj
    securitySchemes:
        bearerAuth:
            scheme: bearer
            bearerFormat: JWT
            type: http
            x-bearerInfoFunc: crosscutting.auth.authentication.authenticate_user
...